added several todos...threading demo not sending msgs consistently.
