+++ src/exo/gait_state_estimator/forceplate/ZMQ_PubSub.py
@@ -1,36 +1,37 @@
 """
-This module implements a publisher/subscriber networking protocol using ZMQ. 
-Its development was motivated by the need to share timestamps between two python programs on different PCs. 
-My current usage is to have the OSL as a publisher of its current log timestamp. 
-A script on a Vicon PC will subscribe to that timestamp to embed sync behavior in a vicon log. 
+This module implements a publisher/subscriber networking protocol using ZMQ.
+Its development was motivated by the need to share timestamps between two python programs on different PCs.
+My current usage is to have the OSL as a publisher of its current log timestamp.
+A script on a Vicon PC will subscribe to that timestamp to embed sync behavior in a vicon log.
 
-See the test_sub() and test_pub() methods for example usages. 
+See the test_sub() and test_pub() methods for example usages.
 
 Requires pyzmq
 
 Kevin Best, 8/17/2023
 """
 import zmq
+from typing import Tuple
 
 class Subscriber():
-    def __init__(self, publisher_ip = 'localhost', publisher_port = "5556", 
+    def __init__(self, publisher_ip = 'localhost', publisher_port = "5556",
                  timeout_ms = 100, topic_filter = '', get_latest_only = True) -> None:
         """
-        Instantiate a subscriber object. Requires the IP address of the publisher, the port of the publisher, an optional timeout flag, optional topic filter, and a bool to ask for only the latest value if there are more than one in the buffer. 
+        Instantiate a subscriber object. Requires the IP address of the publisher, the port of the publisher, an optional timeout flag, optional topic filter, and a bool to ask for only the latest value if there are more than one in the buffer.
         """
         # Socket to talk to server
         context = zmq.Context()
         self.socket = context.socket(zmq.SUB)
         self.socket.setsockopt_string(zmq.SUBSCRIBE, topic_filter)
         if get_latest_only:
-            self.socket.setsockopt(zmq.CONFLATE, 1) 
+            self.socket.setsockopt(zmq.CONFLATE, 1)
         self.socket.connect(("tcp://" + publisher_ip + ":%s") % publisher_port)
         self.timeout_ms = timeout_ms
         self.encoding =  'UTF-8'
-    
-    def get_message(self) -> (str, str, bool):
+
+    def get_message(self) -> Tuple[str, str, bool]:
         """
-        Checks for a message from the subscribed topics. If no message is available in the timeout, it'll return blank topic and message. The msg_received flag will also be false. 
+        Checks for a message from the subscribed topics. If no message is available in the timeout, it'll return blank topic and message. The msg_received flag will also be false.
         Returns topic, message, msg_received flag
         """
         msg_received = self.socket.poll(self.timeout_ms)
@@ -46,8 +47,8 @@ class Subscriber():
 
 
 class Publisher():
-    """ 
-    Instantiates a publisher object. Only required input is a port. Any subscriber on the network can subscribe to this topic via the IP address of the publisher. 
+    """
+    Instantiates a publisher object. Only required input is a port. Any subscriber on the network can subscribe to this topic via the IP address of the publisher.
     """
     def __init__(self, port = "5556") -> None:
         self.context = zmq.Context()
@@ -57,7 +58,7 @@ class Publisher():
 
     def publish(self, topic, message) -> None:
         """
-        Publish a message on a specified topic. Note that topic names CANNOT have spaces in them. 
+        Publish a message on a specified topic. Note that topic names CANNOT have spaces in them.
         """
         assert " " not in topic, "topic name cannot have spaces!"
         self.socket.send_string(topic + " " + message)
@@ -65,7 +66,7 @@ class Publisher():
 
 def testSub():
     """
-    Import this method and run in a thread to test the subscriber behavior. Once you start a publisher, you should see the data parsed. 
+    Import this method and run in a thread to test the subscriber behavior. Once you start a publisher, you should see the data parsed.
     """
     sub = Subscriber()
     while(1):
@@ -75,7 +76,7 @@ def testSub():
 
 def testPub():
     """
-    Import this method and run in a thread to see publisher behavior. You should see whatever data you sent here appear in a subscriber. 
+    Import this method and run in a thread to see publisher behavior. You should see whatever data you sent here appear in a subscriber.
     """
     pub = Publisher()
     i = 0
