--- threading_demo_updated.py
@@ -1,829 +1,0 @@
-# Abstract Base class for threading
-import threading
-import queue
-
-from typing import Type, Dict, Any, Optional
-from abc import ABC, abstractmethod
-from src.utils.flexible_sleeper import FlexibleSleeper
-from non_singleton_logger import NonSingletonLogger
-from dephyEB51 import DephyEB51Actuator
-
-# logging setup
-from src.utils.filing_utils import get_logging_info
-from opensourceleg.logging import Logger, LogLevel
-CONSOLE_LOGGER = Logger(enable_csv_logging=False,
-                        log_path=get_logging_info(user_input_flag=False)[0],
-                        stream_level = LogLevel.INFO,
-                        log_format = "%(levelname)s: %(message)s"
-                        )
-
-from Exoboot_Messenger_Hub import MessageRouter
-
-class BaseWorkerThread(threading.Thread, ABC):
-    """
-    Abstract base class for worker threads in the exoskeleton system.
-    This class provides a template for creating threads that can be paused and quit.
-    It includes methods for thread lifecycle management and a run method that
-    handles the thread's main loop.
-    """
-
-    def __init__(self,
-                 quit_event:Type[threading.Event],
-                 pause_event:Type[threading.Event],
-                 log_event:Type[threading.Event],
-                 name:str,
-                 frequency:int=100)->None:
-
-        super().__init__(name=name)
-        self.quit_event = quit_event    # event to signal quitting
-        self.pause_event = pause_event  # event to signal pausing
-        self.log_event = log_event      # event to signal logging
-
-        self.daemon = True
-        self.frequency = int(frequency)
-        self.rt_loop = FlexibleSleeper(dt=1/frequency)  # create a soft real-time loop with the specified frequency
-
-        # set-up a logger for each thread/instance
-        # logger_name = f"{name}_logger"
-        # log_path = "./src/logs/"
-        # self.data_logger = NonSingletonLogger(
-        #     log_path=log_path,
-        #     file_name=logger_name,
-        #     file_level=logging.DEBUG,
-        #     stream_level=logging.INFO
-        # )
-
-    def run(self)->None:
-        """
-        Main loop structure for each instantiated thread.
-        """
-        LOGGER.debug(f"[{self.name}] Thread running.")
-
-        while self.quit_event.is_set(): # while not quitting
-            self.pre_iterate()              # run a pre-iterate method
-
-            if not self.pause_event.is_set():
-                self.on_pause()             # if paused, run once
-            else:
-                try:
-                    self.iterate()          # call the iterate method to perform the thread's task
-                except Exception as e:
-                    LOGGER.debug(f"Exception in thread {self.name}: {e}")
-
-            self.post_iterate()             # run a post-iterate method
-            self.rt_loop.pause()
-
-        LOGGER.debug(f"[{self.name}] Thread exiting.")
-
-    @abstractmethod
-    def pre_iterate(self):
-        """
-        Override this abstract method in subclasses.
-        This method is called BEFORE the iterate method.
-        It can be used to perform any repetitive setup or initialization tasks.
-        """
-        LOGGER.debug(f"[{self.name}] pre-iterate() called.")
-        pass
-
-    @abstractmethod
-    def iterate(self):
-        """
-        Override this abstract method in subclasses to do the thread's work.
-        This method contains the main logic for the thread's operation.
-        """
-        LOGGER.debug(f"[{self.name}] iterate() called.")
-        pass
-
-    @abstractmethod
-    def post_iterate(self):
-        """
-        Override this abstract method in subclasses.
-        This method is called AFTER the iterate method.
-        It can be used to perform any repetitive cleanup or logging, or finalization tasks.
-        """
-        LOGGER.debug(f"[{self.name}] post-iterate() called.")
-        pass
-
-    @abstractmethod
-    def on_pause(self):
-        """
-        Override this abstract method in subclasses.
-        Runs once before pausing threads.
-        """
-        LOGGER.debug(f"[{self.name}] on_pre_pause() called.")
-        pass
-
-
-class ActuatorThread(BaseWorkerThread):
-    """
-    Threading class for EACH actuator.
-    This class handles the actuator's state updates and manages the actuator's control loop.
-    """
-    def __init__(self,
-                 name: str,
-                 actuator,
-                 msg_router,
-                 quit_event: Type[threading.Event],
-                 pause_event: Type[threading.Event],
-                 log_event: Type[threading.Event],
-                 frequency: int = 100) -> None:
-
-        super().__init__(quit_event, pause_event, log_event, name, frequency=frequency)
-
-        self.actuator = actuator
-        self.actuator.start()  # start the actuator
-
-        self.msg_router = msg_router
-        self.inbox = None
-
-    def pre_iterate(self)->None:
-        # Check mailbox for messages from GSE or other threads
-        mail_list = self.inbox.get_all_mail()
-        for mail in mail_list:
-            self.decode_message(mail)
-
-    def decode_message(self, mail):
-        try:
-            for key, value in mail.contents.items():
-                # TODO: ensure key matches the actuator's name (want to act according to the actuator's side)
-                if hasattr(self, key):
-                    setattr(self, key, value)
-        except Exception as err:
-            self.data_logger.debug(f"Error decoding message: {err}")
-
-    def iterate(self)->None:
-        """
-        Main loop for the actuator thread.
-        This method is called repeatedly in the thread's run loop.
-        It handles the actuator's state updates.
-        """
-        self.actuator.update()
-        CONSOLE_LOGGER.debug(f"motor position: {self.actuator.motor_position:.2f}")
-
-        # TODO: use exoboots context manager to potentially execute code
-            # TODO: command assistance
-
-    def post_iterate(self)->None:
-        # TODO add rest of stuff
-        if self.log_event.is_set():
-            LOGGER.debug(f"[{self.name}] log_event True")
-
-    def on_pause(self)->None:
-        pass
-
-
-class GaitStateEstimatorThread(BaseWorkerThread):
-    """
-    Threading class for the Gait State Estimator.
-    This class handles gait state estimation for BOTH exoskeletons/sides together.
-    """
-    def __init__(self,
-                 msg_router,
-                 quit_event:Type[threading.Event],
-                 pause_event:Type[threading.Event],
-                 log_event: Type[threading.Event],
-                 name:Optional[str] = None,
-                 frequency:int=100)->None:
-
-        super().__init__(quit_event, pause_event, log_event, name=name, frequency=frequency)
-
-        self.msg_router = msg_router
-        self.mailbox = None # will be instantiated
-
-        # instantiate the walking simulator
-        self.walker = WalkingSimulator(stride_period=1.20)
-
-        # TODO: initialize GSE Bertec here
-
-    def pre_iterate(self)->None:
-        pass
-
-    def iterate(self):
-        """
-        Main loop for the actuator thread.
-        This method is called repeatedly in the thread's run loop.
-        It handles the actuator's state updates.
-        """
-
-        self.time_in_stride = self.walker.update_time_in_stride()
-        self.ank_angle = self.walker.update_ank_angle()
-
-        # TODO: send corresponding time in stride (left and right) to the corresponding ACTIVE actuators (left and right)
-        try:
-            if hasattr(self.walker, 'left_stride_period'):
-                self.msg_router.send(
-                    sender=self.name,
-                    recipient="left",
-                    contents={"stride_period": self.walker.left_stride_period}
-                )
-            if hasattr(self.walker, 'right_stride_period'):
-                self.msg_router.send(
-                    sender=self.name,
-                    recipient="right",
-                    contents={"stride_period": self.walker.right_stride_period}
-                )
-        except Exception as e:
-            CONSOLE_LOGGER.debug(f"Error sending stride period: {e}")
-
-        CONSOLE_LOGGER.debug(f"Time in stride: {self.time_in_stride:.3f}s")
-        CONSOLE_LOGGER.debug(f"Ankle angle: {self.ank_angle:.2f} deg")
-
-        # TODO: send gait state to the actuators
-
-    def post_iterate(self)->None:
-        pass
-
-    def on_pause(self)->None:
-        pass
-
-
-import sys
-import select
-class GUICommunication(BaseWorkerThread):
-    """
-    Threading class to simulate GUI communication via gRPC.
-    This class handles GUI communication for BOTH exoskeletons/sides together.
-
-    It constantly monitors for new user input specifying a desired torque setpoint.
-    It then sends these setpoints to the actuators to handle.
-    """
-    def __init__(self,
-                 msg_router,
-                 quit_event:Type[threading.Event],
-                 pause_event:Type[threading.Event],
-                 log_event: Type[threading.Event],
-                 name:Optional[str] = None,
-                 frequency:int=100)->None:
-
-
-        super().__init__(quit_event, pause_event, log_event, name=name, frequency=frequency)
-
-        self.msg_router = msg_router
-        self.inbox = None
-        self.torque_setpoint = 20.0
-
-    def pre_iterate(self)->None:
-        # read mail regardless of paused state
-        mail_list = self.inbox.get_all_mail()  # return list of mail
-        CONSOLE_LOGGER.debug(f"mail received in GUICommunication thread {len(mail_list)}")
-
-        # unpackage the mail
-        for mail in mail_list:
-            self.decode_message(mail)
-            print(f"torque setpoint is:{self.torque_setpoint}")
-
-    def decode_message(self, mail:list) -> Optional[Any]:
-        """
-        Decode a message from the mailbox.
-        """
-        try:
-            for key, value in mail.contents.items():
-                # TODO: value validation (not None or something)
-                setattr(self, key, value)
-
-        except Exception as err:
-            LOGGER.debug(f"Error decoding message: {err}")
-            return None
-
-    def iterate(self):
-        """
-        Non-blocking: Only read user input if available.
-        Allow user to input a new desired torque setpoint.
-        If the user doesn't input a new value, the current setpoint is used.
-        """
-
-        # Put the torque setpoint into the queue for the main thread
-        pass
-
-    def post_iterate(self)->None:
-        pass
-
-    def on_pause(self)->None:
-        pass
-
-
-# Mini Exoboots Robot Class for testing threading
-from opensourceleg.robots.base import RobotBase
-from opensourceleg.sensors.base import SensorBase
-from dephyEB51 import DephyEB51Actuator
-from opensourceleg.logging import LOGGER
-from opensourceleg.actuators.base import CONTROL_MODES
-from opensourceleg.actuators.dephy import DEFAULT_CURRENT_GAINS
-from typing import Union, Dict
-import time
-
-
-class DephyExoboots(RobotBase[DephyEB51Actuator, SensorBase]):
-
-    def start(self) -> None:
-        """
-        Start the Exoskeleton.
-        """
-        super().start()
-
-    def stop(self) -> None:
-        """
-        Stop the Exoskeleton.
-        """
-
-        super().stop()
-
-    def set_actuator_mode_and_gains(self, actuator)-> None:
-        """
-        Call the setup_controller method for all actuators.
-        This method selects current control mode and sets PID gains for each actuator.
-        """
-        actuator.set_control_mode(CONTROL_MODES.CURRENT)
-        LOGGER.info("finished setting control mode")
-
-        actuator.set_current_gains(
-            kp=DEFAULT_CURRENT_GAINS.kp,
-            ki=DEFAULT_CURRENT_GAINS.ki,
-            kd=DEFAULT_CURRENT_GAINS.kd,
-            ff=DEFAULT_CURRENT_GAINS.ff,
-        )
-        LOGGER.info("finished setting gains")
-
-    def update(self) -> None:
-        """
-        Update the exoskeleton.
-        """
-        # print(f"Updating exoskeleton robot: {self.tag}")
-        super().update()
-
-    def spool_belts(self):
-        """
-        Spool the belts of both actuators.
-        This method is called to prepare the actuators for operation.
-        """
-        for actuator in self.actuators.values():
-            actuator.spool_belt()
-            CONSOLE_LOGGER.info(f"finished spooling belt of {actuator.side}")
-
-    def set_to_transparent_mode(self):
-        """
-        Set the exo currents to 0mA.
-        """
-        self.update_current_setpoints(current_inputs=0, asymmetric=False)
-        self.command_currents()
-
-    def detect_active_actuators(self) -> Union[str, list[str]]:
-        """
-        Detect active actuators.
-        Returns a string if only one actuator is active, otherwise a list of strings.
-        """
-        active_sides = list(self.actuators.keys())
-
-        if len(active_sides) == 1:
-            return active_sides[0]
-
-        return active_sides
-
-    def create_current_setpts_dict(self) -> None:
-        """
-        create dictionary of current setpoints (in mA) corresponding to actuator side
-        """
-        self.current_setpoints = {}
-        for actuator in self.actuators.values():
-            self.current_setpoints[actuator.side] = 0.0
-
-        # TODO: generate test to determine if current_setpoints dict has the same keys as the actuators dict
-
-    def update_current_setpoints(self, current_inputs: Union[int, Dict[str, int]], asymmetric:bool=False) -> None:
-        """
-        Directly assign currents to the 'current_setpoints' dictionary for current control.
-
-        If symmetric, the same current value is applied to both sides (with motor sign).
-        If asymmetric, the user must pass a dictionary specifying currents for each side.
-
-        Args:
-            - current: int or dict. If symmetric=False, this should be a dict with 'left' and 'right' keys.
-            - asymmetric: bool. If True, use side-specific currents from the dictionary.
-        """
-        # TODO: ensure that current_inputs matches the number of active sides
-            # TODO: if more than the number of active sides provided, trim to active one only
-            # TODO: handle missing sides
-
-        # TODO: clip current setpoints to below max limit
-
-        if asymmetric:
-            for side, current in current_inputs.items():    # assign different currents for each actuator
-                actuator = getattr(self, side)
-                self.current_setpoints[side] = int(current) * actuator.motor_sign
-        else:
-            for side in self.actuators.keys():              # assign the same current for both actuators
-                actuator = getattr(self, side)
-                self.current_setpoints[side] = int(current_inputs) * actuator.motor_sign
-
-    def convert_torque_to_current_setpoints(self, torque_setpoint: float) -> dict:
-        """
-        Find the appropriate current setpoint for the actuators.
-        This method is called to determine the current setpoint based on the torque setpoint.
-
-        arguments:
-            torque_setpoint: float, the desired torque setpoint in Nm.
-
-
-        returns:
-            current_setpoints:   dict of currents for each active actuator.
-                        key is the side of the actuator (left or right).
-        """
-        for actuator in self.actuators.values():
-            self.current_setpoints[actuator.side] = actuator.torque_to_current(torque_setpoint)
-            CONSOLE_LOGGER.info(f"finished finding current setpoint for {actuator.side}")
-
-            return self.current_setpoints
-
-    def command_currents(self) -> None:
-        """
-        Commands current setpoints to each actuator based on the current_setpoints dictionary.
-        The setpoints can be unique.
-        """
-        # TODO: ensure current_setpoints values are integers, no greater than max current limit, and are not None
-
-        for actuator in self.actuators.values():
-
-            current_setpoint = self.current_setpoints.get(actuator.side)
-
-            if current_setpoint is not None:
-                actuator.set_motor_current(current_setpoint)
-                CONSOLE_LOGGER.info(f"Finished setting current setpoint for {actuator.side}")
-            else:
-                CONSOLE_LOGGER.warning(f"Unknown side '{actuator.side}' and unable to command current. Skipping.")
-
-    def initialize_rt_plots(self) -> list:
-        """
-        Initialize real-time plots for the exoskeleton robot.
-        Naming and plotting is flexible to each active actuator.
-
-        The following time series are plotted:
-        - Current (A)
-        - Temperature (°C)
-        - Ankle Angle (°)
-        - Transmission Ratio
-        - Ankle Torque Setpoint (Nm)
-
-        """
-        # converting actuator dictionary keys to a list
-        active_sides_list = list(self.actuators.keys())
-
-        print("Active actuators:", active_sides_list)
-
-        # pre-slice colors based on the number of active actuators
-        colors = ['r', 'b'][:len(active_sides_list)]
-        if len(active_sides_list) > len(colors):
-            raise ValueError("Not enough unique colors for the number of active actuators.")
-
-        # repeat line styles and widths for each active actuator
-        line_styles = ['-' for _ in active_sides_list]
-        line_widths = [2 for _ in active_sides_list]
-
-        current_plt_config = {'names' : active_sides_list,
-                        'colors' : colors,
-                        'line_style': line_styles,
-                        'title' : "Exo Current (A) vs. Sample",
-                        'ylabel': "Current (A)",
-                        'xlabel': "timestep",
-                        'line_width': line_widths,
-                        'yrange': [0,30]
-                        }
-
-        temp_plt_config = {'names' : active_sides_list,
-                        'colors' : colors,
-                        'line_style': line_styles,
-                        'title' : "Case Temperature (°C) vs. Sample",
-                        'ylabel': "Temperature (°C)",
-                        'xlabel': "timestep",
-                        'line_width': line_widths,
-                        'yrange': [20,60]
-                        }
-
-        in_swing_plt_config = {'names' : active_sides_list,
-                        'colors' : colors,
-                        'line_style': line_styles,
-                        'title' : "Bertec in-swing vs. Sample",
-                        'ylabel': "Bool",
-                        'xlabel': "timestep",
-                        'line_width': line_widths,
-                        'yrange': [0,150]
-                        }
-
-        TR_plt_config = {'names' : active_sides_list,
-                        'colors' : colors,
-                        'line_style': line_styles,
-                        'title' : "TR (°) vs. Sample",
-                        'ylabel': "N",
-                        'xlabel': "timestep",
-                        'line_width': line_widths,
-                        'yrange': [0,20]
-                        }
-
-        imu_plt_config = {'names' : active_sides_list,
-                        'colors' : colors,
-                        'line_style': line_styles,
-                        'title' : "Activations vs. Sample",
-                        'ylabel': "Bool",
-                        'xlabel': "timestep",
-                        'line_width': line_widths,
-                        'yrange': [0,50]
-                        }
-
-        plot_config = [current_plt_config, temp_plt_config, in_swing_plt_config, TR_plt_config, imu_plt_config]
-
-        return plot_config
-
-    def update_rt_plots(self, bertec_swing_flag, imu_activations) -> list:
-        """
-        Updates the real-time plots with current values for:
-        - Current (A)
-        - Temperature (°C)
-        - Bertec In swing
-        - Transmission Ratio
-        - IMU estimator activations
-
-        The data is collected from the exoboots object and returned as a list of arrays.
-        This is done for each active actuator only.
-
-        Returns:
-            plot_data_array: A list of data arrays (for active actuators) for each plot.
-        """
-
-        data_to_plt = []
-
-        for actuator in self.actuators.values():
-            data_to_plt.extend([
-                abs(actuator.motor_current),  # Motor current
-                actuator.case_temperature,    # Case temperature
-                bertec_swing_flag,
-                actuator.gear_ratio,          # Gear ratio
-                imu_activations
-            ])
-
-        return data_to_plt
-
-    def track_variables_for_logging(self, logger: Logger) -> None:
-        """
-        Track variables for each active actuator for logging to a single file
-        """
-
-        for actuator in self.actuators.values():
-            dummy_grpc_value = 5.0
-            dummy_ankle_torque_setpt = 20
-            logger.track_variable(lambda: time.time(), "pitime")
-            logger.track_variable(lambda: dummy_grpc_value, "dollar_value")
-            logger.track_variable(lambda: dummy_ankle_torque_setpt, "torque_setpt_Nm")
-
-            logger.track_variable(lambda: actuator.accelx, f"{actuator._tag}_accelx_mps2")
-            logger.track_variable(lambda: actuator.motor_current, f"{actuator._tag}_current_mA")
-            logger.track_variable(lambda: actuator.motor_position, f"{actuator._tag}_position_rad")
-            logger.track_variable(lambda: actuator.motor_encoder_counts, f"{actuator._tag}_encoder_counts")
-            logger.track_variable(lambda: actuator.case_temperature, f"{actuator._tag}_case_temp_C")
-
-            tracked_vars = logger.get_tracked_variables()
-            print("Tracked variables:", tracked_vars)
-
-    @property
-    def left(self) -> DephyEB51Actuator:
-        try:
-            return self.actuators["left"]
-        except KeyError:
-            CONSOLE_LOGGER.error("Ankle actuator not found. Please check for `left` key in the actuators dictionary.")
-            exit(1)
-
-    @property
-    def right(self) -> DephyEB51Actuator:
-        try:
-            return self.actuators["right"]
-        except KeyError:
-            CONSOLE_LOGGER.error("Ankle actuator not found. Please check for `right` key in the actuators dictionary.")
-            exit(1)
-
-
-
-class ThreadManager:
-    """
-    This class manages thread creation, communication and termination for the exoskeleton system.
-    """
-
-    def __init__(self, msg_router:MessageRouter, actuators:Dict) -> None:
-
-        self.actuators = actuators              # Dictionary of Actuators
-        self.msg_router = msg_router            # MessageRouter class instance
-
-        # create threading events common to all threads
-        self._quit_event = threading.Event()    # Event to signal threads to quit.
-        self._pause_event = threading.Event()   # Event to signal threads to pause.
-        self._log_event = threading.Event()     # Event to signal threads to log
-
-        # initialize thread events
-        self._quit_event.set()      # exo is running
-        self._pause_event.clear()   # exo starts paused
-        self._log_event.clear()     # exo starts not logging
-
-        # initialize dict of threads
-        self._threads = {}
-
-    def start(self) -> None:
-        """
-        Creates the following threads:
-         - for each ACTIVE actuator
-         - for the Gait State Estimator
-         - for the GUI
-        """
-
-        for actuator in self.actuators.values():
-            self.initialize_actuator_thread(actuator)
-
-        # creating 1 thread for the Gait State Estimator
-        self.initialize_GSE_thread()
-
-        # creating 1 thread for GUI communication
-        self.initialize_GUI_thread()
-
-    def start_all_threads(self)->None:
-        """
-        Start all threads in the thread manager.
-        This method is called to start all threads after they have been initialized.
-        """
-
-        for thread in self._threads.values():
-            thread.start()
-            LOGGER.debug(f"Thread {thread.name} started.")
-
-    def stop(self) -> None:
-        """Signal threads to quit and join them"""
-
-        # setting the quit event so all threads recieve the kill signal
-        self._quit_event.set()
-        LOGGER.debug("Setting quit event for all threads.")
-
-    def initialize_actuator_thread(self, actuator: DephyEB51Actuator) -> None:
-        """
-        Create and start a thread for the specified actuator.
-        This method is called to set up the actuator communication thread.
-        """
-
-        actuator_thread = ActuatorThread(actuator=actuator,
-                                         quit_event=self._quit_event,
-                                         pause_event=self._pause_event,
-                                         log_event=self._log_event,
-                                         name=f"{actuator.side}",
-                                         frequency=1,
-                                         msg_router=self.msg_router,
-                                         )
-
-        LOGGER.debug(f"created {actuator.side} actuator thread")
-        self._threads[actuator.side] = actuator_thread
-
-    def initialize_GSE_thread(self) -> None:
-        """
-        Create and start the Gait State Estimator thread.
-        This method is called to set up the GSE communication thread.
-        """
-        # create a FIFO queue with max size for inter-thread communication
-        name = "gse"
-        gse_thread = GaitStateEstimatorThread(quit_event=self._quit_event,
-                                              pause_event=self._pause_event,
-                                              log_event=self._log_event,
-                                              name=name,
-                                              frequency=1,
-                                              msg_router=self.msg_router)
-
-        LOGGER.debug(f"created gse thread")
-        self._threads[name] = gse_thread
-
-    def initialize_GUI_thread(self) -> None:
-        """
-        Create and start the GUI thread for user input.
-        This method is called to set up the GUI communication thread.
-        """
-
-        # create a FIFO queue with max size for inter-thread communication
-        name = "gui"
-        gui_thread = GUICommunication(quit_event=self._quit_event,
-                                      pause_event=self._pause_event,
-                                      log_event=self._log_event,
-                                      name=name,
-                                      frequency=1,
-                                      msg_router=self.msg_router)
-        LOGGER.debug(f"created gui thread")
-        self._threads[name] = gui_thread
-
-    def return_active_threads(self)->list:
-        """
-        Return list of active thread addresses.
-        """
-        return self._threads.values()
-
-    def __enter__(self) -> None:
-        """
-        Context manager enter method.
-        This method is called when the ThreadManager is used in a with statement.
-        """
-        self.start()
-
-    def __exit__(self, exc_type, exc_value, traceback) -> None:
-        """
-        Context manager exit method.
-        This method is called when the ThreadManager is used in a with statement.
-        It stops all threads and cleans up.
-        """
-        self.stop()
-
-    @property
-    def left_exo_queue(self) -> Dict[str, queue.Queue]:
-        """Get the queue for the left actuator"""
-        return self.queues.get("left")
-
-    @property
-    def right_exo_queue(self) -> Dict[str, queue.Queue]:
-        """Get the queue for the right actuator"""
-        return self.queues.get("right")
-
-    @property
-    def gse_queue(self) -> Dict[str, queue.Queue]:
-        """Get the queue for the gait state estimator"""
-        return self.queues.get("gse")
-
-    @property
-    def gui_queue(self) -> Dict[str, queue.Queue]:
-        """Get the queue for the GUI communication"""
-        return self.queues.get("gui")
-
-    @property
-    def left_thread(self) -> Optional[ActuatorThread]:
-        """Get the left actuator thread"""
-        return self._threads.get("left")
-
-    @property
-    def right_thread(self) -> Optional[ActuatorThread]:
-        """Get the right actuator thread"""
-        return self._threads.get("right")
-
-    @property
-    def gse_thread(self) -> Optional[ActuatorThread]:
-        """Get the right actuator thread"""
-        return self._threads.get("gse")
-
-    @property
-    def gui_thread(self) -> Optional[ActuatorThread]:
-        """Get the right actuator thread"""
-        return self._threads.get("gui")
-
-
-# Example main loop
-from opensourceleg.utilities import SoftRealtimeLoop
-from src.utils.actuator_utils import create_actuators
-from src.settings.constants import EXO_SETUP_CONST
-from src.utils.walking_simulator import WalkingSimulator
-
-if __name__ == '__main__':
-
-    # create actuators
-    actuators = create_actuators(gear_ratio=1,
-                                 baud_rate=EXO_SETUP_CONST.BAUD_RATE,
-                                 freq=EXO_SETUP_CONST.FLEXSEA_FREQ,
-                                 debug_level=EXO_SETUP_CONST.LOG_LEVEL)
-
-    # create Exoboots Robot
-    exoboots = DephyExoboots(tag="exoboots",
-                             actuators=actuators,
-                             sensors={})
-
-    # create a message router for inter-thread communication
-    msg_router = MessageRouter()
-
-    # instantiate thread manager
-    system_manager = ThreadManager(msg_router=msg_router, actuators=actuators)
-
-    # instantiate soft real-time clock
-    clock = SoftRealtimeLoop(dt = 1 / 1) # Hz
-
-    with system_manager:
-        # set-up addressbook for the PostOffice & create inboxes for each thread
-        msg_router.setup_addressbook(*system_manager.return_active_threads())
-
-        # start all threads
-        system_manager.start_all_threads()
-
-        for t in clock:
-            try:
-
-                # send mail to the GUI thread
-                msg_router.send(sender="main", recipient="gui", contents={"torque_setpoint": 30.0})
-                print("Main sent")
-
-                # TODO: pass reported values into AssistanceGenerator class -> method is part of Exoboots Robot class
-                    # TODO: determine appropriate torque setpoint given current gait state
-
-                # TODO: send torque setpoints to each corresponding actuator
-                # TODO: determine appropriate current setpoint that matches the torque setpoint -> handled by DephyEB51Actuator class (within each actuator thread)
-                # TODO: command appropriate current setpoint using DephyExoboots class
-
-            except KeyboardInterrupt:
-                print("KeyboardInterrupt received.")
-                break
-
-            except Exception as err:
-                print("Unexpected error in executing main controller:", err)
-                break
